+++ lib/math.c
@@ -51,7 +51,7 @@ double pow(double a, double b)
 ///The seed for random number generation.
 static unsigned long long rand_seed = 0L;
 
-void s_rand(unsigned long seed)
+void s_rand(unsigned long long seed)
 {
     //Apply the initial scramble of the seed.
     rand_seed = (seed ^ SEED_MULTI) & SEED_MASK;
